# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Symetra Lab Backend API",
    "description": "Dokumentasi Resmi REST API untuk Workshop Management System Symetra Lab.\\n\\nMencakup manajemen inventori 3D printing (filamen, mesin cetak, sparepart, komponen hardware, packaging), kalkulasi HPP dan margin berbasis **7 Pos Kas**, integrasi Shopee Open Platform v2, pesanan manual, serta antrean produksi otomatis.",
    "version": "2.0.0",
    "contact": {
      "name": "Symetra Lab Team",
      "url": "https://symetralab.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server (v1 standard)"
    },
    {
      "url": "http://localhost:8080/api/v2",
      "description": "Local Development Server (v2 alias)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan Supabase JWT Access Token"
      },
      "UserIDHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-ID",
        "description": "Header alternatif ID pengguna untuk local development / testing"
      }
    },
    "schemas": {
      "StandardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Operasi berhasil dilakukan"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Deskripsi error"
          },
          "error": {
            "type": "string",
            "example": "Detail teknis error"
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login Pengguna via Supabase",
        "description": "Mengautentikasi pengguna menggunakan email dan password Supabase Auth.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "admin@symetralab.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "rahasia123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Email atau password salah",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Profil Pengguna Aktif",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profil pengguna ditemukan"
          },
          "401": {
            "description": "Token kedaluwarsa atau tidak valid"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Daftar Kategori Produk",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar seluruh kategori produk workshop"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Tambah Kategori Baru",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Drone Parts"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kategori berhasil dibuat"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Ubah Nama Kategori",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kategori berhasil diubah"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Hapus Kategori",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kategori berhasil dihapus"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Daftar Produk & Kalkulasi 7 Pos Kas",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar produk dengan cost breakdown lengkap"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Buat Produk & Resep BOM Baru",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "category",
                  "material_type"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Articulated Dragon Toy"
                  },
                  "category": {
                    "type": "string",
                    "example": "Toys"
                  },
                  "material_type": {
                    "type": "string",
                    "example": "PLA Silk"
                  },
                  "default_weight_grams": {
                    "type": "number",
                    "example": 140.0
                  },
                  "default_print_time_hours": {
                    "type": "number",
                    "example": 4.5
                  },
                  "target_margin_percent": {
                    "type": "integer",
                    "example": 40
                  },
                  "base_selling_price": {
                    "type": "number",
                    "example": 75000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produk berhasil dibuat"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Detail Produk & Komponen BOM",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail produk lengkap"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Perbarui Spesifikasi Produk",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produk berhasil diperbarui"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Hapus Produk",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produk berhasil dihapus"
          }
        }
      }
    },
    "/products/by-sku/{sku}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Cari Produk Berdasarkan SKU",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produk ditemukan"
          }
        }
      }
    },
    "/products/auto-generate-skus": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Otomatisasi Penetapan SKU Baru",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "SKU berhasil dibuat untuk produk yang belum memiliki SKU"
          }
        }
      }
    },
    "/filaments": {
      "get": {
        "tags": [
          "Filaments"
        ],
        "summary": "Daftar Spool Filamen",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar filamen dengan profil teknis preloaded"
          }
        }
      },
      "post": {
        "tags": [
          "Filaments"
        ],
        "summary": "Daftarkan Spool Filamen Baru",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "color_name",
                  "price_per_roll"
                ],
                "properties": {
                  "brand": {
                    "type": "string",
                    "example": "eSUN"
                  },
                  "material_type": {
                    "type": "string",
                    "example": "PLA+"
                  },
                  "color_name": {
                    "type": "string",
                    "example": "Pine Green"
                  },
                  "color_hex": {
                    "type": "string",
                    "example": "#2E5834"
                  },
                  "price_per_roll": {
                    "type": "number",
                    "example": 175000
                  },
                  "current_stock_grams": {
                    "type": "number",
                    "example": 1000
                  },
                  "low_stock_threshold_grams": {
                    "type": "number",
                    "example": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Spool filamen berhasil didaftarkan"
          }
        }
      }
    },
    "/filaments/{id}/stock": {
      "patch": {
        "tags": [
          "Filaments"
        ],
        "summary": "Update / Timbang Fisik Stok Filamen",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "current_stock_grams"
                ],
                "properties": {
                  "current_stock_grams": {
                    "type": "number",
                    "example": 782.5
                  },
                  "is_weighed": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stok filamen berhasil diperbarui"
          }
        }
      }
    },
    "/machines": {
      "get": {
        "tags": [
          "Machines"
        ],
        "summary": "Daftar Armada 3D Printer",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar printer dan suku cadang maintenance"
          }
        }
      },
      "post": {
        "tags": [
          "Machines"
        ],
        "summary": "Daftarkan Printer Baru",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "total_purchase_cost"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Bambu Lab P1S"
                  },
                  "brand": {
                    "type": "string",
                    "example": "Bambu Lab"
                  },
                  "total_purchase_cost": {
                    "type": "number",
                    "example": 11500000
                  },
                  "lifespan_hours": {
                    "type": "integer",
                    "example": 10000
                  },
                  "avg_power_watts": {
                    "type": "integer",
                    "example": 220
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Printer berhasil ditambahkan"
          }
        }
      }
    },
    "/machines/{id}/state": {
      "patch": {
        "tags": [
          "Machines"
        ],
        "summary": "Ubah Status Operasional Mesin",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "state"
                ],
                "properties": {
                  "state": {
                    "type": "string",
                    "enum": [
                      "IDLE",
                      "PRINTING",
                      "MAINTENANCE",
                      "OFFLINE"
                    ],
                    "example": "PRINTING"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status mesin diperbarui"
          }
        }
      }
    },
    "/machines/{id}/parts": {
      "get": {
        "tags": [
          "Machines"
        ],
        "summary": "Daftar Suku Cadang Perawatan Printer",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar sparepart printer"
          }
        }
      },
      "post": {
        "tags": [
          "Machines"
        ],
        "summary": "Tambah Suku Cadang Mesin",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "part_name",
                  "lifespan_hours"
                ],
                "properties": {
                  "part_name": {
                    "type": "string",
                    "example": "Hardened Steel Nozzle 0.4"
                  },
                  "lifespan_hours": {
                    "type": "number",
                    "example": 1500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sparepart berhasil ditambahkan"
          }
        }
      }
    },
    "/machines/parts/{part_id}/replace": {
      "post": {
        "tags": [
          "Machines"
        ],
        "summary": "Reset Jam Pakai Sparepart (Setelah Diganti)",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jam operasional suku cadang di-reset ke 0"
          }
        }
      }
    },
    "/components": {
      "get": {
        "tags": [
          "Components"
        ],
        "summary": "Daftar Komponen Hardware",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar komponen hardware (baut, bearing)"
          }
        }
      },
      "post": {
        "tags": [
          "Components"
        ],
        "summary": "Tambah Komponen Hardware",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price_per_unit"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Stainless Steel Bolt M3x12"
                  },
                  "price_per_unit": {
                    "type": "number",
                    "example": 350
                  },
                  "markup_percent": {
                    "type": "number",
                    "example": 20
                  },
                  "description": {
                    "type": "string",
                    "example": "DIN 912 Hex Socket"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Komponen berhasil dibuat"
          }
        }
      }
    },
    "/packaging-items": {
      "get": {
        "tags": [
          "Packaging"
        ],
        "summary": "Daftar Bahan Kemasan Individual",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar kardus, bubble wrap, polymailer"
          }
        }
      }
    },
    "/packaging-presets": {
      "get": {
        "tags": [
          "Packaging"
        ],
        "summary": "Daftar Bundel Preset Kemasan",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar preset packaging"
          }
        }
      }
    },
...
    "/orders/{id}/status": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "Ubah Status Pengerjaan Pesanan",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "IN_PRODUCTION",
                      "COMPLETED",
                      "DELIVERED",
                      "CANCELLED"
                    ],
                    "example": "IN_PRODUCTION"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status pesanan berhasil diperbarui"
          }
        }
      }
    },
...
    "/shopee/orders": {
      "get": {
        "tags": [
          "Shopee"
        ],
        "summary": "Daftar Pesanan Shopee Online",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "UserIDHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar pesanan Shopee beserta breakdown keuangan 7 Pos Kas"
          }
        }
      }
    },
...
  }
}