# internal/delivery/http/router.go
package http

import (
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/labstack/echo/v4/middleware"

	"symetra-lab-backend-v2/config"
	"symetra-lab-backend-v2/internal/delivery/http/handler"
	appMiddleware "symetra-lab-backend-v2/internal/delivery/http/middleware"
)

type RouterConfig struct {
	AppConfig        *config.Config
	AuthHandler      *handler.AuthHandler
	ProductHandler   *handler.ProductHandler
	CategoryHandler  *handler.CategoryHandler
	OrderHandler     *handler.OrderHandler
	FilamentHandler  *handler.FilamentHandler
	MachineHandler   *handler.MachineHandler
	ComponentHandler *handler.ComponentHandler
	PackagingHandler *handler.PackagingHandler
	ShopeeHandler    *handler.ShopeeHandler
	ShopConfigHandler *handler.ShopConfigHandler
	ProductionHandler *handler.ProductionHandler
}

func SetupRouter(e *echo.Echo, cfg RouterConfig) {
	// Global Middlewares
	e.Use(middleware.Logger())
	e.Use(middleware.Recover())
	e.Use(appMiddleware.SetupCORS())

	// Health check
	e.GET("/health", func(c echo.Context) error {
		return c.JSON(http.StatusOK, map[string]string{
			"status":  "healthy",
			"version": "2.0.0",
		})
	})

	e.GET("/", func(c echo.Context) error {
		return c.JSON(http.StatusOK, map[string]string{
			"app":     "Symetra Lab Backend v2",
			"version": "2.0.0",
			"status":  "online",
		})
	})

	// OpenAPI Specification & Interactive Documentation UI (Scalar)
	e.File("/openapi.json", "docs/openapi.json")

	scalarHTML := `<!doctype html>
<html>
  <head>
    <title>Symetra Lab API Reference</title>
    <meta charset="utf-8" />
    <meta name="viewport" content="width=device-width, initial-scale=1" />
    <link rel="icon" type="image/svg+xml" href="https://scalar.com/favicon.svg" />
    <style>
      body { margin: 0; font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif; }
    </style>
  </head>
  <body>
    <script
      id="api-reference"
      data-url="/openapi.json"
      data-configuration='{"theme": "purple", "darkMode": true, "showSidebar": true}'
      src="https://cdn.jsdelivr.net/npm/@scalar/api-reference"></script>
  </body>
</html>`

	e.GET("/docs", func(c echo.Context) error {
		return c.HTML(http.StatusOK, scalarHTML)
	})

	e.GET("/swagger", func(c echo.Context) error {
		return c.Redirect(http.StatusMovedPermanently, "/docs")
	})


	// Mount routes for both /api/v2 and /api/v1 (dual compatibility for frontend)
	apiPrefixes := []string{"/api/v2", "/api/v1"}
	for _, prefix := range apiPrefixes {
		api := e.Group(prefix)

		// Public Shopee OAuth & Push Webhook routes
		if cfg.ShopeeHandler != nil {
			api.GET("/shopee/auth-url", cfg.ShopeeHandler.GetAuthURL)
			api.GET("/shopee/callback", cfg.ShopeeHandler.HandleCallback)
			api.POST("/shopee/webhook", cfg.ShopeeHandler.HandlePushWebhook)
			api.POST("/shopee/push-listener", cfg.ShopeeHandler.HandlePushWebhook)
		}

	// Public Auth routes
	if cfg.AuthHandler != nil {
		auth := api.Group("/auth")
		auth.POST("/login", cfg.AuthHandler.Login)
		auth.POST("/refresh", cfg.AuthHandler.RefreshToken)
	}

	// Protected routes (require Supabase JWT Bearer token or X-User-ID header)
	authMiddleware := appMiddleware.AuthMiddleware(cfg.AppConfig)
	protected := api.Group("", authMiddleware)

	if cfg.AuthHandler != nil {
		protected.GET("/auth/me", cfg.AuthHandler.GetProfile)
		protected.POST("/auth/logout", cfg.AuthHandler.Logout)
	}

	// Categories routes
	if cfg.CategoryHandler != nil {
		categories := protected.Group("/categories")
		categories.GET("", cfg.CategoryHandler.List)
		categories.POST("", cfg.CategoryHandler.Create)
		categories.PUT("/:id", cfg.CategoryHandler.Update)
		categories.DELETE("/:id", cfg.CategoryHandler.Delete)
	}

	// Products routes
	if cfg.ProductHandler != nil {
		products := protected.Group("/products")
		products.GET("", cfg.ProductHandler.List)
		products.GET("/by-sku/:sku", cfg.ProductHandler.GetBySKU)
		products.GET("/:id", cfg.ProductHandler.GetByID)
		products.POST("", cfg.ProductHandler.Create)
		products.PUT("/:id", cfg.ProductHandler.Update)
		products.DELETE("/:id", cfg.ProductHandler.Delete)
		products.POST("/generate-skus", cfg.ProductHandler.AutoGenerateDraftSKUs)
		products.POST("/auto-generate-skus", cfg.ProductHandler.AutoGenerateDraftSKUs)
	}

	// Orders routes (Manual/Direct Orders Hub)
	if cfg.OrderHandler != nil {
		orders := protected.Group("/orders")
		orders.GET("", cfg.OrderHandler.List)
		orders.GET("/:id", cfg.OrderHandler.GetByID)
		orders.POST("", cfg.OrderHandler.Create)
		orders.PATCH("/:id/status", cfg.OrderHandler.UpdateStatus)
		orders.PATCH("/:id/payment", cfg.OrderHandler.UpdatePaymentStatus)
		orders.DELETE("/:id", cfg.OrderHandler.Delete)
	}

	// Filaments & Profiles & Material Rates
	if cfg.FilamentHandler != nil {
		filaments := protected.Group("/filaments")
		filaments.GET("", cfg.FilamentHandler.ListFilaments)
		filaments.GET("/:id", cfg.FilamentHandler.GetFilamentByID)
		filaments.POST("", cfg.FilamentHandler.Create)
		filaments.PUT("/:id", cfg.FilamentHandler.Update)
		filaments.PATCH("/:id/stock", cfg.FilamentHandler.SyncStock)
		filaments.POST("/:id/sync-stock", cfg.FilamentHandler.SyncStock)
		filaments.DELETE("/:id", cfg.FilamentHandler.Delete)

		protected.GET("/filament-profiles", cfg.FilamentHandler.ListFilamentProfiles)

		rates := protected.Group("/filament-material-rates")
		rates.GET("", cfg.FilamentHandler.ListMaterialRates)
		rates.POST("", cfg.FilamentHandler.CreateMaterialRate)
		rates.PUT("/:id", cfg.FilamentHandler.UpdateMaterialRate)
		rates.DELETE("/:id", cfg.FilamentHandler.DeleteMaterialRate)
	}

	// Machines & Maintenance Parts
	if cfg.MachineHandler != nil {
		machines := protected.Group("/machines")
		machines.GET("", cfg.MachineHandler.ListMachines)
		machines.GET("/:id", cfg.MachineHandler.GetMachineByID)
		machines.POST("", cfg.MachineHandler.Create)
		machines.PUT("/:id", cfg.MachineHandler.Update)
		machines.PATCH("/:id/state", cfg.MachineHandler.UpdateState)
		machines.DELETE("/:id", cfg.MachineHandler.Delete)

		machines.GET("/:id/parts", cfg.MachineHandler.GetPartsByMachine)
		machines.POST("/:id/parts", cfg.MachineHandler.CreatePart)
		machines.PUT("/parts/:part_id", cfg.MachineHandler.UpdatePart)
		machines.POST("/parts/:part_id/replace", cfg.MachineHandler.ReplacePart)
		machines.DELETE("/parts/:part_id", cfg.MachineHandler.DeletePart)
	}

	// Components
	if cfg.ComponentHandler != nil {
		components := protected.Group("/components")
		components.GET("", cfg.ComponentHandler.ListComponents)
		components.GET("/:id", cfg.ComponentHandler.GetComponentByID)
		components.POST("", cfg.ComponentHandler.Create)
		components.PUT("/:id", cfg.ComponentHandler.Update)
		components.DELETE("/:id", cfg.ComponentHandler.Delete)
	}

	// Packaging Items & Presets
	if cfg.PackagingHandler != nil {
		pkgItems := protected.Group("/packaging-items")
		pkgItems.GET("", cfg.PackagingHandler.ListPackagingItems)
		pkgItems.GET("/:id", cfg.PackagingHandler.GetPackagingItemByID)
		pkgItems.POST("", cfg.PackagingHandler.CreatePackagingItem)
		pkgItems.PUT("/:id", cfg.PackagingHandler.UpdatePackagingItem)
		pkgItems.DELETE("/:id", cfg.PackagingHandler.DeletePackagingItem)

		pkgPresets := protected.Group("/packaging-presets")
		pkgPresets.GET("", cfg.PackagingHandler.ListPackagingPresets)
		pkgPresets.POST("", cfg.PackagingHandler.CreatePackagingPreset)
		pkgPresets.PUT("/:id", cfg.PackagingHandler.UpdatePackagingPreset)
		pkgPresets.DELETE("/:id", cfg.PackagingHandler.DeletePackagingPreset)
	}
	// Shopee Integration & 7 Pos Kas Finance
	if cfg.ShopeeHandler != nil {
		shopeeGroup := protected.Group("/shopee")
		shopeeGroup.GET("/shops", cfg.ShopeeHandler.ListShops)
		shopeeGroup.POST("/shops/:shop_id/refresh", cfg.ShopeeHandler.RefreshToken)
		shopeeGroup.POST("/sync-shopee", cfg.ShopeeHandler.SyncShopee)
		shopeeGroup.GET("/orders", cfg.ShopeeHandler.ListOrders)
		shopeeGroup.GET("/orders/:order_sn", cfg.ShopeeHandler.GetOrderDetail)
		shopeeGroup.POST("/link-sku", cfg.ShopeeHandler.LinkSKU)
		shopeeGroup.POST("/orders/:order_sn/ship", cfg.ShopeeHandler.ShipOrder)
		shopeeGroup.GET("/orders/:order_sn/shipping-label", cfg.ShopeeHandler.DownloadShippingLabel)
		shopeeGroup.GET("/financial/cash-summary", cfg.ShopeeHandler.GetCashflowSummary)
		shopeeGroup.GET("/financial/cashflow-summary", cfg.ShopeeHandler.GetCashflowSummary)
		shopeeGroup.POST("/financial/recalculate", cfg.ShopeeHandler.RecalculateFinances)
	}

	// Config (Shop & Marketplaces)
	if cfg.ShopConfigHandler != nil {
		configGroup := protected.Group("/config")
		configGroup.GET("/shop", cfg.ShopConfigHandler.GetShopConfig)
		configGroup.PUT("/shop", cfg.ShopConfigHandler.UpdateShopConfig)
		configGroup.GET("/marketplaces", cfg.ShopConfigHandler.ListMarketplaces)
		configGroup.POST("/marketplaces", cfg.ShopConfigHandler.CreateMarketplace)
		configGroup.PUT("/marketplaces/:id", cfg.ShopConfigHandler.UpdateMarketplace)
		configGroup.DELETE("/marketplaces/:id", cfg.ShopConfigHandler.DeleteMarketplace)
	}


	// Production Queue & Auto-Deduct
	if cfg.ProductionHandler != nil {
		prodGroup := protected.Group("/production")
		prodGroup.GET("/queue", cfg.ProductionHandler.GetQueue)
		prodGroup.POST("/complete-job", cfg.ProductionHandler.CompleteJob)
	}

	}
}